Check for a specific xsiType when adding a disk to a VMware instance
https://bugzilla.redhat.com/show_bug.cgi?id=1311221

When hot adding a disk we were using a Regex to find the
appropriate SCSI controller.  This change validates the xsiType
as being the correct type and removes the Regex.
This will remove the problems the Regex
created with different localizations.

(transferred from ManageIQ/manageiq-gems-pending@b95ef35962e8186db915877667af3106bdda6c0a)