Allow the MiqVimBroker to expose a DRb connection as a service
This change allows the DRb server created for the vim broker to be
exposed to clients which are not local.

This will be used when the broker is running in a container which
exposes an openshift service.

This change also configures clients to use the vim-broker service
when running in a container