Add dirty hacks to get DRb to send objects back with the correct uri
This needs to be done because the DRb protocol code assumes that the
server is bound to the same uri that clients use to connect.

This bug manifests itself when you run a DRb server behind a proxy.
The server can't bind to the proxy uri so it uses 0.0.0.0 and the client
can't respond using 0.0.0.0, but DRb sends objects back to the client
with an ip of 0.0.0.0 which raises an error when clients attempt to
call a method on the returned object.

To "fix" this I set the server protocol uri to the service host
which forces the response objects to have the "correct" uri