Remove dependency on gems-pending
Remove the gems-pending dependency which was there to pull in the
util/extensions/miq-openuri extension.  This extension was deleted a
while ago by manageiq-gems-pending@37a63ac4555fc92fe7d8144c92c9f96ba0d2c47c

Tested VM smartstate of a VM with a snapshot (which is the only caller
of MiqVimDataStore#get_file_data) and confirmed that it works correctly
and returns the .vmsd file content.