Allow running the monitorUpdates thread in MiqVim
Now that some of the previous "broker" functionality is being run not
over DRb but from a shared session in a worker we need to be able to run
the monitor updates thread which keeps the inventory cache up to date
from MiqVim not just DMiqVim.

DMiqVim still has some extra logic specifically in error handling that
is specific to being run from within the broker so some of that is still
needed in DMiqVim but the bulk of it can move to MiqVim.